VideoGame_Discounts v1.00
Thanks to the help of JSoup, an open source Java library, the VideoGame_Discounts program can successfully retrieve and parse HTMLs from various websites to read and store data. The data is then displayed in a window created using Java's Swing. In the case of the VideoGame_Discounts program, the data includes deals and discounts in video games, and occasionally video game accessories.